Move source code out of resources folder
