+++ core-plugin/src/com/google/gct/idea/appengine/run/AppEngineRunConfigurationType.java
@@ -15,15 +15,16 @@
  */
 package com.google.gct.idea.appengine.run;
 
-import com.intellij.execution.BeforeRunTask;
-import com.intellij.openapi.util.Key;
-import icons.GoogleCloudToolsIcons;
+import com.google.gct.idea.ui.GoogleCloudToolsIcons;
 
+import com.intellij.execution.BeforeRunTask;
 import com.intellij.execution.configurations.ConfigurationFactory;
 import com.intellij.execution.configurations.ConfigurationType;
 import com.intellij.execution.configurations.ConfigurationTypeUtil;
 import com.intellij.execution.configurations.RunConfiguration;
 import com.intellij.openapi.project.Project;
+import com.intellij.openapi.util.Key;
+
 import org.jetbrains.annotations.NotNull;
 
 import javax.swing.Icon;
